requirements.txt
openai==0.28.1
vosk==0.3.45
sounddevice==0.4.6
pyttsx3==2.90
python-dotenv==1.0.1
